ZED-31 add implementation
