+++ backend/queries/recipes/update_ingredient_in_recipe.sql
@@ -1,0 +1,4 @@
+UPDATE ingredients_recipes
+SET amount = $3
+WHERE ingredient_id = $2
+AND recipe_id = $1
